Rename notificatons to index
